fix: Context api를 jotai로 변경
